refactor(spells): Refactored spell list and replaced getSpellLevelDisplayText with an array
