+++ src/utils.js
@@ -1,14 +1,68 @@
-export const getSpellLevelDisplayText = spellLevel => {
-  switch (Number(spellLevel)) {
-    case 0:
-      return 'Cantrip'
-    case 1:
-      return '1st level'
-    case 2:
-      return '2nd level'
-    case 3:
-      return '3rd level'
-    default:
-      return `${spellLevel}th level`
+export const spellLevels = [
+  {
+    value: 0,
+    label: 'Cantrip'
+  },
+  {
+    value: 1,
+    label: '1st level'
+  },
+  {
+    value: 2,
+    label: '2nd level'
+  },
+  {
+    value: 3,
+    label: '3rd level'
+  },
+  {
+    value: 4,
+    label: '4th level'
+  },
+  {
+    value: 5,
+    label: '5th level'
+  },
+  {
+    value: 6,
+    label: '6th level'
+  },
+  {
+    value: 7,
+    label: '7th level'
+  },
+  {
+    value: 8,
+    label: '8th level'
+  },
+  {
+    value: 9,
+    label: '9th level'
   }
-}
+]
+
+export const spellSchools = [
+  'Abjuration',
+  'Conjuration',
+  'Divination',
+  'Enchantment',
+  'Evocation',
+  'Illusion',
+  'Necromancy',
+  'Transmutation'
+]
+
+export const spellComponents = [
+  {
+    value: 'V',
+    label: 'Verbal (V)'
+  },
+  {
+    value: 'M',
+    label: 'Material (M)'
+  },
+  {
+    value: 'S',
+    label: 'Somatic (S)'
+  }
+]
